Conflicts:
	graph/graph.h
	shared/gl/fbo.cpp
	shared/gl/fbo.h
	shared/gl/texture.cpp
	sociarium/agent.cpp
	sociarium/common.cpp
	sociarium/common.h
	sociarium/cvframe.cpp
	sociarium/designtide.cpp
	sociarium/font.cpp
	sociarium/glew.cpp
	sociarium/language.cpp
	sociarium/main.cpp
	sociarium/menu_and_message.h
	sociarium/module/community_detection.cpp
	sociarium/module/community_detection.h
	sociarium/module/community_detection_information_flow_mapping_using_greedy_method.cpp
	sociarium/module/graph_creation.cpp
	sociarium/module/graph_creation.h
	sociarium/module/graph_creation_read_adjacency_matrix.cpp
	sociarium/module/language_en.cpp
	sociarium/module/layout.cpp
	sociarium/module/layout.h
	sociarium/sociarium_graph.cpp
	sociarium/sociarium_graph.h
	sociarium/sociarium_graph_time_series.cpp
	sociarium/texture.cpp
	sociarium/thread/detail/read_file.cpp
	sociarium/thread/graph_creation.cpp
	sociarium/thread/graph_retouch.cpp
	sociarium/world.h
	sociarium/world_impl.h
	sociarium/world_impl_draw.cpp
	sociarium/world_impl_draw_orthogonal_part.cpp
	sociarium/world_impl_edit_graph.cpp
	sociarium/world_impl_initialize.cpp
	sociarium/world_impl_mouse_action.cpp
	sociarium/world_impl_others.cpp
	sociarium/world_impl_select.cpp
	sociarium/world_impl_thread.cpp
	sociarium/world_impl_view.cpp

+++ sociarium/Classes/SociariumView.mm
@@ -28,9 +28,11 @@ using namespace hashimoto_ut;
 - (void)prepareOpenGL
 {
   [super prepareOpenGL];
+
+  if (world_)
+    World::destroy(world_);
   
-  World::destroy(world_);
-  world_ = World::create(static_cast<CGLContextObj>([[self openGLContext] CGLContextObj]));
+  world_ = World::create([self window], static_cast<CGLContextObj>([[self openGLContext] CGLContextObj]));
   
   if (redrawTimer)
   {
